Extract `Bridge.Deposit.test.ts`
